+++ +sbfsem/+util/sem.m
@@ -1,0 +1,4 @@
+function sem = sem(data)
+  % find standard error of the mean
+  sem = std(data)/sqrt(length(data));
+end